fix: error in rev_list
The rev_list function was reversing the list but duplicating the first
element of the initial list which would be present at the beginning and
the end of the final list. The one present at the beginning final list
is removed and it now works perfectly.

+++ c-lib/tests/lib/linked_list/test_my_rev_list.c
@@ -21,7 +21,7 @@ Test(my_rev_list, rev_linked_list)
     linked_list = start;
     my_rev_list(&linked_list);
     for (int i = 9; i >= 0; i--) {
-        cr_assert_eq(linked_list->data, i * 21);
+        cr_assert_eq((size_t)linked_list->data, i * 21);
         linked_list = linked_list->next;
     }
 }
